requirements.txt
Django==2.0.5
django-adminlte2==0.3.0
django-registration-redux==2.4
djangorestframework==3.8.2
Pillow==5.1.0
pkg-resources==0.0.0
PyMySQL==0.8.0
pytz==2018.4
